refactor

+++ AudioProcessing/SpectrumAnalyzer.h
@@ -56,7 +56,7 @@ namespace AudioProcessing
 
 		for (size_t outIndex=0;outIndex < outputSize;outIndex++)
 		{
-			int inValueIntIndex = floor(fInputIndex);
+			int inValueIntIndex = (int) floor(fInputIndex);
 
 			if (fNextIndex - fInputIndex < 1.0)
 			{
@@ -83,7 +83,7 @@ namespace AudioProcessing
 				}
 				// Add end values
 				sum += pInput[inValueIntIndex] * (1.f - fInputIndex + inValueIntIndex);
-				if (inValueIntNextIndex < inputSize)
+				if (inValueIntNextIndex < (int) inputSize)
 				{
 					sum += pInput[inValueIntNextIndex] * (fNextIndex - inValueIntNextIndex);
 				}
